% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20066R.R
\name{indep_norm}
\alias{indep_norm}
\title{A Independence sampler using R}
\usage{
indep_norm(m, a, b, p, n, mu, sigma)
}
\arguments{
\item{m}{the length of chain}

\item{a}{parameter of Beta(a,b) proposal dist.}

\item{b}{parameter of Beta(a,b) proposal dist.}

\item{p}{mixing parameter}

\item{n}{sample size}

\item{mu}{parameters of the normal densities include two numbers}

\item{sigma}{parameters of the normal densities include two numbers}
}
\value{
the estimation of mix parameter p
}
\description{
A Independence sampler using R
}
\examples{
\dontrun{
I<-indep_norm(10000,1,1,0.2,50,c(0, 5),c(1, 1))
print(I$xt)
print(I$p_hat)
plot(xt, type="l", ylab="p")
hist(xt[101:m], main="", xlab="p", prob=TRUE)
}
}
