% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20066R.R
\name{MetropolisRa}
\alias{MetropolisRa}
\title{A Metropolis-Hastings sampler using R}
\usage{
MetropolisRa(m, sigma)
}
\arguments{
\item{m}{the number of samples}

\item{sigma}{the parameter of Rayleigh distribution}
}
\value{
a random sample of size and the number of reject
}
\description{
A Metropolis-Hastings sampler using R
}
\examples{
\dontrun{
MH.Ra <- MetropolisRa(10000,4)
}
}
