% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20066R.R
\name{Ra}
\alias{Ra}
\title{The Rayleigh density function}
\usage{
Ra(x, sigma)
}
\arguments{
\item{x}{the variable}

\item{sigma}{the parameter of Rayleigh density function}
}
\value{
The Rayleigh density function
}
\description{
The Rayleigh density function
}
\examples{
\dontrun{
ra <- Ra(x,4)
}
}
