% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwcc}
\alias{rwcc}
\title{A Metropolis sampler using Rcpp}
\usage{
rwcc(sigma, x0, N)
}
\arguments{
\item{sigma}{the parameter of proposal distribution Normal}

\item{x0}{the initial value of sample}

\item{N}{the number of samples}
}
\value{
a random sample of size N and the number of reject
}
\description{
A Metropolis sampler using Rcpp to generate random variables from a standard Cauchy distribution
}
\examples{
\dontrun{
rwC <- rwcc(0.5,25,2000)
print(rwC$x)
print(rwC$k)
}
}
